# api/openapi.yaml
openapi: 3.1.0
info:
  title: Super CDN Control Plane API
  version: 0.4.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    First formal contract for the stable Super CDN operator API.
    This file describes the current v0.4.0-era behavior for the CLI-facing
    control plane: auth, diagnostics, asset buckets, site deployments, edge
    manifests, and conservative garbage collection.

    Root-only provider provisioning, legacy low-level object repair, and init
    endpoints remain documented in docs/cli-reference.md until their contracts
    are frozen.
servers:
  - url: http://127.0.0.1:8080
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  "/api/v1/asset-buckets/{slug}/objects/primary-target":
    post:
      tags: [Asset Buckets]
      operationId: switchAssetBucketObjectPrimaryTarget
      summary: Switch one bucket object's primary delivery target after explicit confirmation
      description: Refuses buckets that use routing_policy, because primary_target metadata would not control policy-based live routing.
      parameters:
        - $ref: "#/components/parameters/BucketSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrimaryTargetSwitchRequest"
      responses:
        "200":
          description: Switch plan or applied result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrimaryTargetSwitchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  "/api/v1/sites/{id}/deployments/{deployment}":
    get:
      tags: [Deployments]
      operationId: getSiteDeployment
      summary: Read a site deployment
      parameters:
        - $ref: "#/components/parameters/SiteID"
        - $ref: "#/components/parameters/DeploymentID"
      responses:
        "200":
          description: Deployment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteDeployment"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Deployments]
      operationId: deleteSiteDeployment
      summary: Delete a non-active, unpinned deployment
      parameters:
        - $ref: "#/components/parameters/SiteID"
        - $ref: "#/components/parameters/DeploymentID"
        - name: delete_objects
          in: query
          schema:
            type: boolean
            default: false
        - name: delete_remote
          in: query
          schema:
            type: boolean
            default: true
        - name: dry_run
          in: query
          description: Return a deletion safety plan without deleting metadata or remote objects.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Deletion result, or a dry-run safety plan when dry_run=true.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteDeploymentDeleteResult"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SuperCDN token
  parameters:
    BucketSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
    DeploymentID:
      name: deployment
      in: path
      required: true
      schema:
        type: string
    SiteID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource was not found or is not visible to the token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource state prevents the requested operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    DoctorCheck:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
        status:
          type: string
          enum: [ok, warning, error, skipped]
        message:
          type: string
        error:
          type: string
...
    PrimaryTargetSwitchRequest:
      type: object
      required: [path, target]
      properties:
        path:
          type: string
          description: Bucket logical path or site request path.
        target:
          type: string
          description: Ready replica target to make primary.
        expected_current_target:
          type: string
          description: Optional guard that must match the current primary target.
        dry_run:
          type: boolean
          default: true
          description: Defaults to true; set false only with confirm=switch.
        confirm:
          type: string
          description: Must be switch when dry_run is false.
    PrimaryTargetSwitchResponse:
      type: object
      required: [status, mode, resource, path, object_id, previous_target, target, dry_run, effective_now]
      properties:
        status:
          type: string
          enum: [planned, switched, noop]
        mode:
          type: string
          enum: [bucket, site]
        resource:
          type: string
        deployment_id:
          type: string
        path:
          type: string
        file:
          type: string
        object_id:
          type: integer
          format: int64
        previous_target:
          type: string
        target:
          type: string
        dry_run:
          type: boolean
        effective_now:
          type: boolean
        target_url:
          type: string
        target_url_redacted:
          type: boolean
        checks:
          type: array
          items:
            $ref: "#/components/schemas/DoctorCheck"
        warnings:
          type: array
          items:
            type: string
        next_commands:
          type: array
          items:
            type: string
        rollback_command:
          type: string
...
    SiteDeployment:
      type: object
      properties:
        id:
          type: string
        site_id:
          type: string
        environment:
          type: string
        status:
          type: string
        active:
          type: boolean
        route_profile:
          type: string
        deployment_target:
          type: string
        routing_policy:
          type: string
        resource_failover:
          type: boolean
        version:
          type: string
        preview_url:
          type: string
        production_url:
          type: string
        pinned:
          type: boolean
        expires_at:
          type: string
          format: date-time
      additionalProperties: true
    SiteDeploymentDeleteResult:
      type: object
      properties:
        site_id:
          type: string
        deployment_id:
          type: string
        status:
          type: string
          description: Present on dry-run plans.
        deployment_target:
          type: string
          description: Present on dry-run plans.
        active:
          type: boolean
          description: Present on dry-run plans.
        pinned:
          type: boolean
          description: Present on dry-run plans.
        dry_run:
          type: boolean
        delete_objects:
          type: boolean
        delete_remote:
          type: boolean
        safe_to_run:
          type: boolean
          description: Present on dry-run plans; false for active or pinned deployments.
        remote_cleanup_supported:
          type: boolean
          description: Whether the requested delete flags can clean remote resources without an external provider-specific cleanup flow.
        remote_cleanup_blockers:
          type: array
          items:
            type: string
          description: Reasons remote cleanup is not supported by this metadata delete path.
        deleted:
          type: boolean
        deleted_deployment:
          type: boolean
        object_count:
          type: integer
        objects:
          type: array
          items:
            type: object
            additionalProperties: true
        evidence:
          type: object
          properties:
            file_count:
              type: integer
            artifact_sha256:
              type: string
            manifest_key:
              type: string
            cloudflare_static:
              $ref: "#/components/schemas/CloudflareStaticDeploymentEvidence"
        warning:
          type: string
        warnings:
          type: array
          items:
            type: string
        errors:
          type: array
          items:
            type: string
    CloudflareStaticDeploymentEvidence:
      type: object
      properties:
        worker_name:
          type: string
        version_id:
          type: string
        domains:
          type: array
          items:
            type: string
        urls:
          type: array
          items:
            type: string
        compatibility_date:
          type: string
        assets_sha256:
          type: string
        cache_policy:
          type: string
        headers_generated:
          type: boolean
        not_found_handling:
          type: string
        verification_status:
          type: string
        verified_at:
          type: string
        published_at:
          type: string
...